Update what-is-a-state-machine.md
These are mostly preferences on my part; the one I feel most strongly about is the point _A state can define *actions*..._ IMHO the original text implies action definition as a consequence of entering or exiting. I think we want to be very clear that actions may _occur_ on entry/exit, but are defined at an entirely different place and time.